aws_saa_study/requirements.txt
pyautogui==0.9.54
pyperclip==1.8.2
pillow==10.1.0
opencv-python==4.8.1.78
pytesseract==0.3.10
numpy==1.24.3 